requirements.txt

python-decouple==3.0

pymongo==3.4.0

requests==2.18.1

beautifulsoup4==4.6.0